+++ custom_types/bip39/index.d.ts
@@ -1,3 +1,0 @@
-// types not available
-declare module 'bip39';
-declare module "bip39/wordlists/english.json";