+++ packages/iov-bns-governance/types/governor.d.ts
@@ -5,6 +5,7 @@ import {
   ElectionRule,
   Electorate,
   Proposal,
+  Vote,
   VoteOption,
   VoteTx,
 } from "@iov/bns";
@@ -26,6 +27,7 @@ export declare class Governor {
   getElectionRules(electorateId: number): Promise<readonly ElectionRule[]>;
   getElectionRuleById(electionRuleId: number): Promise<ElectionRule>;
   getProposals(): Promise<readonly Proposal[]>;
+  getVotes(): Promise<readonly Vote[]>;
   buildCreateProposalTx(options: ProposalOptions): Promise<CreateProposalTx & WithCreator>;
   buildVoteTx(proposalId: number, selection: VoteOption): Promise<VoteTx & WithCreator>;
 }